Redefine PyThread_ident_t as a size_t, and also use it.

+++ Include/internal/pycore_pythread.h
@@ -116,7 +116,7 @@ PyAPI_FUNC(PyLockStatus) PyThread_acquire_lock_timed_with_retries(
     PyThread_type_lock,
     PY_TIMEOUT_T microseconds);
 
-typedef unsigned long long PyThread_ident_t;
+typedef size_t PyThread_ident_t;
 typedef Py_uintptr_t PyThread_handle_t;
 
 #define PY_FORMAT_THREAD_IDENT_T "llu"
